requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-genai==1.57.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.15.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
prettytable==3.17.0
proto-plus==1.27.0
protobuf==6.33.4
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyecharts==2.0.9
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.52.2
streamlit-analytics==0.4.1
streamlit-echarts==0.4.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
websockets==15.0.1
